ComplianceQaPipeline/pyproject.toml
[project]
name = "complianceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-identity-broker>=0.3.1",
    "azure-search-documents>=11.6.0",
    "azure-monitor-opentelemetry>=1.8.7",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.135.2",
    "firecrawl-py>=4.21.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
    "yt-dlp>=2026.3.17",
]
